Reorganize file-structure for base classes
