+++ src/diagnostics/base/index.ts
@@ -1,0 +1,5 @@
+// -------
+// CLASSES
+// -------
+
+export * from './XMLDiagnostics';
